Break configuration from the setup recipe
The setup recipe included a number of resources that are intended to be
run only once. However, the configuration of the baremetal nodes, as it
is an extension to nova.con needs to be present for all nova services.
And these services may or may not be colocated on the same node. So,
make configuration its own recipe.